market cap float
adjusted market cap data type to flaot

+++ schema.sql
@@ -139,7 +139,7 @@ CREATE TABLE Exchange_ETFs (
 CREATE TABLE Cryptocurrency (
     Cryptocurrency_Symbol VARCHAR(10),
     total_supply LONG INT NOT NULL,
-    market_cap DOUBLE NOT NULL,
+    market_cap FLOAT NOT NULL,
     unit_price FLOAT NOT NULL,
     cryptocurrency_name VARCHAR(255) NOT NULL,
     PRIMARY KEY (Cryptocurrency_Symbol)
